// frontend/src/services/api/model/serviceRequest.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * The Calling API
 * Central API contract for The Calling backend and frontend service integration.
 * OpenAPI spec version: 1.1.0
 */
import type { ServiceRequestPriority } from './serviceRequestPriority';
import type { ServiceRequestStatus } from './serviceRequestStatus';

export interface ServiceRequest {
  assignee_id?: string;
  created_at?: string;
  creator_id?: string;
  description?: string;
  id?: string;
  priority?: ServiceRequestPriority;
  status?: ServiceRequestStatus;
  title?: string;
  updated_at?: string;
}