// frontend/src/services/api/model/updateRequest.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * The Calling API
 * Central API contract for The Calling backend and frontend service integration.
 * OpenAPI spec version: 1.1.0
 */
import type { UpdateRequestPriority } from './updateRequestPriority';
import type { UpdateRequestStatus } from './updateRequestStatus';

export interface UpdateRequest {
  assignee_id?: string;
  description?: string;
  priority?: UpdateRequestPriority;
  status?: UpdateRequestStatus;
  title?: string;
}